Adjustments for gcc compiler (mac os)

+++ COLLADAMaya/include/COLLADAMayaEffectExporter.h
@@ -86,7 +86,7 @@ namespace COLLADAMaya
         struct ExportedEffect
         {
             MObject* shader; // The shader object, that was use to export the effect
-            DWORD mColor;  // The color that was used to export the effect
+            unsigned long mColor;  // The color that was used to export the effect
         };
 
         /** Pointer to the document exporter */
